refactor: moved check_disk_vendor to lshw
